requirements.txt
fastapi
uvicorn
pydantic
opencv-python
mediapipe
numpy
requests
python-dotenv
